backend/requirements.txt
# Core Flask
Flask>=3.0.0
Flask-CORS>=4.0.0
Werkzeug>=3.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# PDF processing
PyMuPDF>=1.23.0

# Excel processing
openpyxl>=3.1.0

# Image processing (puede requerir compilación en Windows)
Pillow>=10.0.0

# QR Code processing (opcional, puede fallar en Windows sin dependencias del sistema)
# pyzbar>=0.1.9  # Descomentar si lo necesitas y tienes las dependencias instaladas

# Firebase
firebase-admin>=6.0.0
google-cloud-firestore>=2.0.0

# Utilities
python-dotenv>=1.0.0
